Uploading Part 1
This part is the currently "functional" version of the code all it does is implement the Lehmer32(&mut self) -> u32; method into the update loop and test it out. The code here is as close as I can get this phase of the tutorial to the original C++ code, but the white pixel density is far higher than what is shown in the video. To get roughly the same density of generated pixels this code:

b_is_star = self.lehmer_32() % 256 < 32; // Same as video code wise, but white pixel density is higher

Needs to be modified so instead of the modulo being 256 it is ~2048:

b_is_star = self.lehmer_32() % 2048 < 32; // ~10x the modulo divisor as in the video, but the density of white pixels here is closer to what is shown