Adding part 2 (READ DESC!)
This is the in progress part of the tutorial. In this part of the C++ version a class is added to the program for StarSystem - I am having a lot of trouble figuring out how to implement all this together since the Lehmer32() is referenced by other functions that need to be added, but it is external to the StarSystem implementation code. I think there is a fundamental concept I'm missing for how to properly create something equivalent to a very interconnected C++ class in rust using traits, structs, and impl.